feat: rename strategy files
