+++ anti-fraud/src/anti-fraud/strategy/strategy.selector.ts
@@ -1,4 +1,4 @@
-import {ValidationStrategy} from "./ValidationStrategy";
+import {ValidationStrategy} from "./validation.strategy";
 import {TransactionDto} from "../dto/transaction.dto";
 
 export class StrategySelector {
